Vote Revealer + Spectator List update
Vote revealer now shows who voted yes or no. You can't see what vote it is for
and if it passed or not. But it's a beginning

Spectator list can now also show global spectator list, this will show
who is spectating who, and not just who is spectating the local player.