Add clear damage list button
This resets the damage list. Users are supposed to do this every
time they enter a game. If someone knows how to do it with an event
please open a PR, I'd rather not do it by checking if localPlayer is
in game, because if you decide to rejoin the damage list will be
reset.